# Optimized runtime speed and updated development dependencies