--- a/package.json
+++ b/package.json
@@ -34,15 +34,15 @@
   "devDependencies": {
     "codecov.io": "^0.1.6",
     "console-group": "^0.3.3",
-    "eslint": "^3.19.0",
-    "eslint-plugin-import": "^2.2.0",
+    "eslint": "^5.6.0",
+    "eslint-plugin-import": "^2.14.0",
     "istanbul": "^0.4.5",
-    "mocha": "^3.2.0",
-    "remap-istanbul": "^0.9.5",
-    "rollup": "^0.54.0",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-typescript": "^0.8.1",
-    "typescript": "^2.7.1"
+    "mocha": "^5.2.0",
+    "remap-istanbul": "^0.12.0",
+    "rollup": "^0.66.0",
+    "rollup-plugin-node-resolve": "^3.4.0",
+    "rollup-plugin-typescript": "^1.0.0",
+    "typescript": "^3.0.3"
   },
   "files": [
     "dist/*.js",