Hi!

I was looking at Rollup build performance and found this module consumes a lot of memory and CPU time. I couldn't optimize the memory consumption and only changes that had any impact on performance are on this PR.

1 - I cached `mappings.length` to `len` variable (.5s diff)
2 - I changed type of array to Int8Array which had biggest impact (4-5s diff)

Total time of rollup build in our application:
```
Before:
67.107s
67.535s
68.158s
68.301s
70.307s

After:
63.797s
65.111s
65.029s
65.675s
63.914s
```