@Havunen 2018-09-20
What I tried:
Changed if else to switch
Inlined charsToInteger 
Tried moving allocation of empty segment inside if branch, which had negative impact for some strange reason.