@Rich-Harris 2018-09-20
Fantastic, thank you! Released 1.4.2