@Havunen 2018-09-20
I noticed some tests fail in rollup repository if it updates this dependency, I don't know yet why those tests fail. Maybe it has something to do with the type change of Segment