@Havunen 2018-09-20
Seems like Int16 passes the tests, I need to investigate more what is the root cause and fix it / PR test here