@Havunen 2018-09-20
I have identified the issue and will open new PR with added test case